feat(ot): switch overtime to Double across DTO/entity/mapper, add repo upsert+targeted delete, make service save idempotent (dedupe+upsert), support fractional allocatedHours, fix MySQL 1062 duplicates.

+++ src/main/java/life/work/IntFit/backend/dto/daily_assignments/OvertimeDTO.java
@@ -16,5 +16,5 @@ public class OvertimeDTO {
     private Long teamMemberId;
 
     /** The overtime hours worked in addition to the default 8h day. */
-    private int overtimeHours;
+    private Double overtimeHours; // was int
 }
